backend/pyproject.toml
[project]
name = "resume-qa-backend"
version = "0.1.0"
description = "智能简历解析与流式问答POC"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.9",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "elasticsearch>=8.17,<9.0",
    "redis[hiredis]>=5.0",
    "dashscope>=1.20",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "chardet>=5.2",
    "Pillow>=10.0",
    "PyJWT>=2.8",
    "openpyxl>=3.1",
    "mammoth (>=1.11.0,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
]
